requirements.txt
directord>=0.10.0

pbr!=2.1.0,>=2.0.0
networkx
taskflow
pyyaml
stevedore
ansible_runner
